Extract tableheader
